% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_next.R
\name{get_next}
\alias{get_next}
\title{Gets the next beacon response to the given timestamp}
\usage{
get_next(ts)
}
\arguments{
\item{ts}{timestamp as a unix or POSIXct/POSIXlt value}
}
\value{
A NISTBeaconResponse object
}
\description{
Gets the next beacon response to the given timestamp
}
\examples{
ts <- as.POSIXct("2014-01-01 12:00:00", tz="UTC")
get_next(ts)
}

