% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_record.R
\name{get_record}
\alias{get_record}
\title{Gets the NIST beacon record for a given timestamp}
\usage{
get_record(ts)
}
\arguments{
\item{ts}{timestamp as a unix or POSIXct/POSIXlt value}
}
\value{
A NISTBeaconResponse object
}
\description{
Gets the NIST beacon record for a given timestamp
}
\examples{
ts <- as.POSIXct("2014-01-01 12:00:00", tz="UTC")
get_record(ts)
}

