% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Generic function that does the calling and processes the beacon's response}
\usage{
get_response(type, ts = TS_NOW)
}
\arguments{
\item{type}{Indicates which of the beacon services is called}

\item{ts}{The timestamp to be used, by default the current one}
}
\value{
A NISTBeaconResponse object
}
\description{
This functions is intended to be called by the wrapper functions
}

