% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_start_chain.R
\name{get_start_chain}
\alias{get_start_chain}
\title{Gets the start of the chain for the given timestamp}
\usage{
get_start_chain(ts)
}
\arguments{
\item{ts}{timestamp as a unix or POSIXct/POSIXlt value}
}
\value{
A NISTBeaconResponse object
}
\description{
Gets the start of the chain for the given timestamp
}
\examples{
ts <- as.POSIXct(Sys.time(), tz="UTC")
get_start_chain(ts)
}

