% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NISTBeaconResponse.R
\name{NISTBeaconResponse}
\alias{NISTBeaconResponse}
\title{A class representing the NIST Randomness Beacon response}
\usage{
NISTBeaconResponse(response)
}
\arguments{
\item{response}{The raw XML response from the service}
}
\value{
A NISTBeaconResponse object
}
\description{
A class representing the NIST Randomness Beacon response
}

