% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nrb_get_last.r
\name{nrb_get_last}
\alias{nrb_get_last}
\title{Gets the last beacon response}
\usage{
nrb_get_last()
}
\value{
An nrb_response object
}
\description{
Gets the last beacon response
}
\examples{
nrb_get_last()
}

