% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrb_get_next.r
\name{nrb_get_next}
\alias{nrb_get_next}
\title{Gets the next beacon response to the given timestamp}
\usage{
nrb_get_next(ts)
}
\arguments{
\item{ts}{timestamp as a unix or POSIXct/POSIXlt value}
}
\value{
An nrb_response object
}
\description{
Gets the next beacon response to the given timestamp
}
\examples{
ts <- as.POSIXct("2014-01-01 12:00:00", tz="UTC")
nrb_get_next(ts)
}

