% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nrb_get_start_chain.r
\name{nrb_get_start_chain}
\alias{nrb_get_start_chain}
\title{Gets the start of the chain for the given timestamp}
\usage{
nrb_get_start_chain(ts)
}
\arguments{
\item{ts}{timestamp as a unix or POSIXct/POSIXlt value}
}
\value{
An nrb_response object
}
\description{
Gets the start of the chain for the given timestamp
}
\examples{
ts <- as.POSIXct(Sys.time(), tz="UTC")
nrb_get_start_chain(ts)
}

