% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrb_response.r
\name{nrb_response}
\alias{nrb_response}
\title{A class representing the NIST Randomness Beacon response}
\usage{
nrb_response(response, endpoint)
}
\arguments{
\item{response}{The raw XML response from the service}

\item{endpoint}{The REST endpoint for the service call}
}
\value{
An nrb_response object
}
\description{
A class representing the NIST Randomness Beacon response
}

