% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_NISTBeaconResponse.R
\name{print.NISTBeaconResponse}
\alias{print.NISTBeaconResponse}
\title{Pretty prints a beacon response}
\usage{
\method{print}{NISTBeaconResponse}(response, ...)
}
\arguments{
\item{response}{a NISTBeaconResponse object}
}
\value{
A string representation of the response
}
\description{
Pretty prints a beacon response
}
\examples{
response <- get_last()
print(response)
}

