% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nrb_response.r
\name{print.nrb_response}
\alias{print.nrb_response}
\title{Pretty prints a beacon response}
\usage{
\method{print}{nrb_response}(x, ...)
}
\arguments{
\item{x}{a nrb_response object}

\item{...}{passed to generic print function}
}
\value{
A string representation of the response
}
\description{
Pretty prints a beacon response
}
\examples{
response <- nrb_get_last()
print(response)
}

