% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNISTRandomnessBeacon.R
\docType{package}
\name{RNISTRandomnessBeacon}
\alias{RNISTRandomnessBeacon}
\alias{RNISTRandomnessBeacon-package}
\title{RNISTRandomnessBeacon: A package to access the NIST Randomness Beacon Service}
\description{
This package implements interfaces to access the NIST Randomness Beacon
through its REST web service API. The beacon is currently at
https://beacon.nist.gov/home, and its impelementation is documented at
http://www.nist.gov/itl/csd/ct/nist_beacon.cfm
}

