% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nrb_response.r
\name{nrb_response}
\alias{nrb_response}
\title{A class representing the NIST Randomness Beacon response}
\usage{
nrb_response(response)
}
\arguments{
\item{response}{The raw XML response from the service}
}
\value{
An nrb_response object
}
\description{
A class representing the NIST Randomness Beacon response
}

